requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
easyocr==1.7.1
pillow==10.1.0
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1
